DEV-1880_move_styles_to_customize_route
