+++ spa/src/components/dashboard/sidebar/DashboardSidebar.js
@@ -1,5 +1,5 @@
 import * as S from './DashboardSidebar.styled';
-import { DONATIONS_SLUG, CONTENT_SLUG } from 'routes';
+import { DONATIONS_SLUG, CONTENT_SLUG, STYLES_SLUG } from 'routes';
 import { ICONS } from 'assets/icons/SvgIcon';
 
 import logout from 'components/authentication/logout';
@@ -32,6 +32,16 @@ function DashboardSidebar({ shouldAllowDashboard }) {
             disabled={!shouldAllowDashboard}
           >
             Content
+          </S.NavItem>
+        ) : null}
+        {hasContentSectionAccess ? (
+          <S.NavItem
+            data-testid="nav-content-item"
+            to={STYLES_SLUG}
+            onClick={handleClick}
+            disabled={!shouldAllowDashboard}
+          >
+            Customize
           </S.NavItem>
         ) : null}
         {hasContributionsSectionAccess ? (
