+++ spa/src/routes.js
@@ -2,6 +2,7 @@ export const LOGIN = '/login';
 export const DASHBOARD_SLUG = '/dashboard';
 export const DONATIONS_SLUG = DASHBOARD_SLUG + '/contributions';
 export const CONTENT_SLUG = DASHBOARD_SLUG + '/content';
+export const STYLES_SLUG = DASHBOARD_SLUG + '/styles';
 export const CONNECT_SLUG = DASHBOARD_SLUG + '/connect';
 
 export const EDITOR_ROUTE = DASHBOARD_SLUG + '/edit';
